+++ user_code/bsp/bsp_fdcan.h
@@ -6,7 +6,7 @@
 void can_bsp_init(void);
 void can_filter_init(void);
 uint8_t fdcanx_send_data(FDCAN_HandleTypeDef *hfdcan, uint16_t id, uint8_t *data, uint32_t len);
-uint8_t fdcanx_receive(FDCAN_HandleTypeDef *hfdcan, uint8_t *buf);
+uint16_t fdcanx_receive(FDCAN_HandleTypeDef *hfdcan, uint8_t *buf);
 uint32_t get_fdcan_dlc(uint32_t len);
 __weak void fdcan1_rx_callback(void);
 __weak void fdcan2_rx_callback(void);
